Update README.md
This commit adds a README file for the Superstore SQL Queries project. The README file provides a detailed description of the project and how to use it. It includes the following sections:

Installation: This section explains how to install SSMS and the sample data set required to run the SQL code.
Usage: This section explains how to create and populate the table with the sample data set and how to execute the queries in SSMS.
License: This section specifies the MIT License for the project and provides a link to the LICENSE file.
The README file uses Markdown syntax to format the content and make it more readable and attractive. It also uses third-party services to generate badges that show the status and quality of the project.

This commit aims to improve the documentation and presentation of the project and make it easier for users and collaborators to understand and use it.